dope

+++ shell/dotfiles_install.sh
@@ -23,30 +23,40 @@ if [[ "$OS_TYPE" == "Linux" ]]; then
     ln -s ~/github/notes/configs/dotfiles/.bashrc ~/.bashrc
     # Source the new .bashrc
     source ~/.bashrc
+    echo "Dotfiles installation complete. Please restart your terminal."
+
     # Exit after Linux setup
     exit 0
 fi
 
-# First delete any existing dotfiles in the home directory.
-rm -rf ~/.zshrc ~/.zprofile ~/.gitconfig
+# macOS specific setup
+elseif [[ "$OS_TYPE" == "macOS" ]]; then
+    echo "Performing macOS specific setup..."
 
-# Link the dotfiles from the notes repository to the home directory.
-ln -s ~/github/notes/configs/dotfiles/.zprofile ~/.zprofile
-ln -s ~/github/notes/configs/dotfiles/.zshrc ~/.zshrc
-ln -s ~/github/notes/configs/dotfiles/.gitconfig ~/.gitconfig
+    # First delete any existing dotfiles in the home directory.
+    rm -rf ~/.zshrc ~/.zprofile ~/.gitconfig
 
-# Link python to python3 from homebrew.
-if [[ "$(uname -m)" == "arm64" ]]; then
-    rm -rf /opt/homebrew/bin/python /opt/homebrew/bin/pip
-    ln -s /opt/homebrew/bin/python3 /opt/homebrew/bin/python
-    ln -s /opt/homebrew/bin/pip3 /opt/homebrew/bin/pip
-else
-    rm -rf /usr/local/bin/python /usr/local/bin/pip
-    ln -s /usr/local/bin/python3 /usr/local/bin/python
-    ln -s /usr/local/bin/pip3 /usr/local/bin/pip
+    # Link the dotfiles from the notes repository to the home directory.
+    ln -s ~/github/notes/configs/dotfiles/.zprofile ~/.zprofile
+    ln -s ~/github/notes/configs/dotfiles/.zshrc ~/.zshrc
+    ln -s ~/github/notes/configs/dotfiles/.gitconfig ~/.gitconfig
+
+    # Link python to python3 from homebrew.
+    if [[ "$(uname -m)" == "arm64" ]]; then
+        rm -rf /opt/homebrew/bin/python /opt/homebrew/bin/pip
+        ln -s /opt/homebrew/bin/python3 /opt/homebrew/bin/python
+        ln -s /opt/homebrew/bin/pip3 /opt/homebrew/bin/pip
+    else
+        rm -rf /usr/local/bin/python /usr/local/bin/pip
+        ln -s /usr/local/bin/python3 /usr/local/bin/python
+        ln -s /usr/local/bin/pip3 /usr/local/bin/pip
+    fi
+
+    # Source the new files
+    source ~/.zprofile
+    source ~/.zshrc
+
+    echo "Dotfiles installation complete. Please restart your terminal."
+    exit 0
 fi
-
-# Source the new files
-source ~/.zprofile
-source ~/.zshrc
